Cache tweaks

+++ lib/multilingual/translation/file.rb
@@ -42,7 +42,7 @@ class ::Multilingual::TranslationFile
   
   def remove
     if exists?
-      File.delete(path)
+      File.delete(path) if File.exist?(path)
       after_remove
     end
   end
@@ -58,7 +58,6 @@ class ::Multilingual::TranslationFile
   end
   
   def after_all
-    Multilingual::Language.refresh!
     Multilingual::Translation.refresh!
     Multilingual::Language.refresh!
     Multilingual.refresh_clients(@code)
